Update PumpsIndustry to new design
