+++ src/main/java/zabbum/oelremake/Pumps/PumpsOperations.java
@@ -1,0 +1,80 @@
+package zabbum.oelremake.Pumps;
+
+import com.googlecode.lanterna.TextColor;
+import com.googlecode.lanterna.gui2.Window;
+import lombok.Getter;
+import zabbum.oelremake.AbstractIndustry;
+import zabbum.oelremake.Drills.DrillsIndustry;
+import zabbum.oelremake.Oilfield;
+import zabbum.oelremake.Operations.BuyableIndustryOperation;
+import zabbum.oelremake.Operations.BuyableProductsOperation;
+import zabbum.oelremake.Operations.SabotableOperation;
+import zabbum.oelremake.Player;
+
+import java.util.Map;
+
+@Getter
+public class PumpsOperations implements BuyableIndustryOperation, BuyableProductsOperation, SabotableOperation {
+
+    private final PumpsIndustry[] industries;
+
+    public PumpsOperations() { industries = initialize(); }
+
+    /**
+     * Initialize drills industries.
+     *
+     * @return array of cars industries
+     */
+    public static PumpsIndustry[] initialize() {
+        PumpsIndustry[] pumpProds = new PumpsIndustry[2];
+
+        // Pump productions initialization
+        pumpProds[0] = new PumpsIndustry("ZASSANICKI GMBH");
+        pumpProds[1] = new PumpsIndustry("DR PUMPENER");
+
+        return pumpProds;
+    }
+
+    @Override
+    public void buyIndustryMenu(Player player, Window window, Map<String, String> langMap) throws InterruptedException {
+        BuyableIndustryOperation.defaultBuyIndustryMenu(
+                player,
+                industries,
+                TextColor.ANSI.WHITE_BRIGHT,
+                TextColor.ANSI.BLUE,
+                TextColor.ANSI.BLUE,
+                TextColor.ANSI.WHITE_BRIGHT,
+                TextColor.ANSI.WHITE_BRIGHT,
+                TextColor.ANSI.CYAN,
+                TextColor.ANSI.BLUE,
+                langMap.get("pumpsIndustrySale"),
+                langMap.get("pumpsIndustryPrompt"),
+                langMap.get("pumpsPricePrompt"),
+                50000,
+                window,
+                langMap
+        );
+    }
+
+    @Override
+    public void buyProductsMenu(Player player, Oilfield[] oilfields, Window window, Map<String, String> langMap) throws InterruptedException {
+        BuyableProductsOperation.defaultBuyProductsMenu(
+                player,
+                industries,
+                oilfields,
+                TextColor.ANSI.WHITE_BRIGHT,
+                TextColor.ANSI.BLACK_BRIGHT,
+                TextColor.ANSI.BLACK_BRIGHT,
+                TextColor.ANSI.WHITE_BRIGHT,
+                TextColor.ANSI.WHITE_BRIGHT,
+                TextColor.ANSI.CYAN,
+                TextColor.ANSI.BLACK_BRIGHT,
+                TextColor.ANSI.BLACK,
+                langMap.get("pumpsHereYouCanBuy"),
+                langMap.get("pumpsProductsAmountPrompt"),
+                15,
+                window,
+                langMap
+        );
+    }
+}
