--- src/main/java/zabbum/oelremake/PumpsIndustry.java
@@ -1,68 +1,0 @@
-package zabbum.oelremake;
-
-import com.googlecode.lanterna.TextColor;
-
-import java.util.Random;
-
-public class PumpsIndustry extends AbstractIndustry {
-
-    // Constructor
-    public PumpsIndustry(String name) {
-        super(name);
-
-        Random random = new Random();
-
-        this.setIndustryPrice(random.nextInt(80000) + 36000);
-        this.setProductsAmount(((int) (this.getIndustryPrice() / 10000)) * 7 + 25);
-    }
-
-    public static PumpsIndustry[] initialize() {
-        PumpsIndustry[] pumpProds = new PumpsIndustry[2];
-
-        // Pump productions initialization
-        pumpProds[0] = new PumpsIndustry("ZASSANICKI GMBH");
-        pumpProds[1] = new PumpsIndustry("DR PUMPENER");
-
-        return pumpProds;
-    }
-
-    // Menu for buying pumps productions
-    public static void buyIndustry(Player player, GameProperties gameProperties)
-            throws InterruptedException {
-        AbstractIndustry.buyIndustry(
-                player,
-                gameProperties,
-                PumpsIndustry.class,
-                TextColor.ANSI.WHITE_BRIGHT,
-                TextColor.ANSI.BLUE,
-                TextColor.ANSI.BLUE,
-                TextColor.ANSI.WHITE_BRIGHT,
-                TextColor.ANSI.WHITE_BRIGHT,
-                TextColor.ANSI.CYAN,
-                TextColor.ANSI.BLUE,
-                gameProperties.langMap.get("pumpsIndustrySale"),
-                gameProperties.langMap.get("pumpsIndustryPrompt"),
-                gameProperties.langMap.get("pumpsPricePrompt"),
-                50000);
-    }
-
-    // Menu for buying pumps
-    public static void buyProduct(Player player, GameProperties gameProperties)
-            throws InterruptedException {
-        AbstractIndustry.buyProduct(
-                player,
-                gameProperties,
-                PumpsIndustry.class,
-                TextColor.ANSI.WHITE_BRIGHT,
-                TextColor.ANSI.BLACK_BRIGHT,
-                TextColor.ANSI.BLACK_BRIGHT,
-                TextColor.ANSI.WHITE_BRIGHT,
-                TextColor.ANSI.WHITE_BRIGHT,
-                TextColor.ANSI.CYAN,
-                TextColor.ANSI.BLACK_BRIGHT,
-                TextColor.ANSI.BLACK,
-                gameProperties.langMap.get("pumpsHereYouCanBuy"),
-                gameProperties.langMap.get("pumpsProductsAmountPrompt"),
-                15);
-    }
-}
